Initial work on a Window component
It can have width or height specified, even though the unit system
is not handled well at the monent. And they are just stacked on
top of each other.

It can also take a title property, which can be a string or other
component. The nested components inside will be the content of the
window.

Rearranged `ReactUI` component to display all components in a
structured way. Using icon fonts for visual effect.